refactor: reduce unnecessary file structure
