+++ extension/src/entrypoints/background/listeners/onCommand.ts
@@ -1,7 +1,7 @@
 import type { Command } from "@@/wxt.config";
 import { match } from "ts-pattern";
 import { DisplayMode, ExtStorage } from "@/constants";
-import { sendMessage } from "@/messaging/message";
+import { sendMessage } from "@/message";
 import { getGeneralSettings, setGeneralSettings } from "@/storage/settings";
 
 export const registerOnCommand = () => {
