+++ extension/src/entrypoints/background/listeners/onOpenOptions.ts
@@ -1,4 +1,4 @@
-import { onMessage } from "@/messaging/message";
+import { onMessage } from "@/message";
 
 export const registerOnOpenOptionsMessage = () => {
   onMessage("openOptionsPage", () => {
